Refactored components and modules into separate files
