+++ src/main/scala/play/api/cache/redis/package.scala
@@ -12,6 +12,5 @@ package object redis extends AnyRef with util.Expiration {
   type AsynchronousResult[ A ] = scala.concurrent.Future[ A ]
 
   private[ redis ] type RedisInstance = configuration.RedisInstance
-  private[ redis ] type RedisInstanceManager = configuration.RedisInstanceManager
   private[ redis ] type RedisConnector = connector.RedisConnector
 }
